# eakte-api/work-in-progress/api-doc 2026-08-28.yaml
openapi: 3.0.0
info:
  title: DMS REST API
  version: 1.2.8-snapshot
servers:
  - url: /api
...
paths:
...
  '/incomings/{objaddress}':
    get:
      tags:
        - /incomings
      summary: Eingang lesen
      operationId: ReadIncoming
      parameters:
        - name: x-anwendung
          in: header
          description: Aufrufende Fachanwendung
          required: false
          schema:
            type: string
          example: SWH
        - name: userlogin
          in: header
          description: >-
            Übergabe des Benutzer-Logins, von deren Berechtigung die Aktion
            durchgeführt werden soll.
          required: false
          schema:
            type: string
          example: max.mustermann
        - name: joboe
          in: header
          description: COO-ID der Organisationseinheit des Nutzers
          required: false
          schema:
            type: string
          example: COO.1.2300.1.847
        - name: jobposition
          in: header
          description: 'Rolle, in der der Nutzer handelt'
          required: false
          schema:
            type: string
          example: DocumentManager
        - name: objaddress
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadIncomingAntwortDTO'
        '400':
          description: Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten.
        '401':
          description: Falscher oder fehlender technischer Nutzer.
        '500':
          description: Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten.
      security:
        - basicAuthSecurityDefintion: []
    put:
      tags:
        - /incomings
      summary: Eingang ändern
      operationId: UpdateIncoming
      parameters:
        - name: x-anwendung
          in: header
          description: Aufrufende Fachanwendung
          required: false
          schema:
            type: string
          example: SWH
        - name: userlogin
          in: header
          description: >-
            Übergabe des Benutzer-Logins, von deren Berechtigung die Aktion
            durchgeführt werden soll.
          required: false
          schema:
            type: string
          example: max.mustermann
        - name: joboe
          in: header
          description: COO-ID der Organisationseinheit des Nutzers
          required: false
          schema:
            type: string
          example: COO.1.2300.1.847
        - name: jobposition
          in: header
          description: 'Rolle, in der der Nutzer handelt'
          required: false
          schema:
            type: string
          example: DocumentManager
        - name: objaddress
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateIncomingRequestBodyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateIncomingAntwortDTO'
        '400':
          description: Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten.
        '401':
          description: Falscher oder fehlender technischer Nutzer.
        '500':
          description: Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten.
      security:
        - basicAuthSecurityDefintion: []
  /incomings:
    post:
      tags:
        - /incomings
      summary: Eingang anlegen
      operationId: CreateIncoming
      parameters:
        - name: x-anwendung
          in: header
          description: Aufrufende Fachanwendung
          required: false
          schema:
            type: string
          example: SWH
        - name: userlogin
          in: header
          description: >-
            Übergabe des Benutzer-Logins, von deren Berechtigung die Aktion
            durchgeführt werden soll.
          required: false
          schema:
            type: string
          example: max.mustermann
        - name: joboe
          in: header
          description: COO-ID der Organisationseinheit des Nutzers
          required: false
          schema:
            type: string
          example: COO.1.2300.1.847
        - name: jobposition
          in: header
          description: 'Rolle, in der der Nutzer handelt'
          required: false
          schema:
            type: string
          example: DocumentManager
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateIncomingRequestBodyDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIncomingAntwortDTO'
        '400':
          description: Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten.
        '401':
          description: Falscher oder fehlender technischer Nutzer.
        '500':
          description: Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten.
      security:
        - basicAuthSecurityDefintion: []
...
components:
  schemas:
...
    ArrayOfstring:
      type: object
      properties:
        string:
          type: array
          items:
            type: string
...
    CreateIncomingRequestBodyDTO:
      required:
        - params
      type: object
      properties:
        params:
          $ref: '#/components/schemas/CreateIncomingBasisAnfrageDTO'
        giattachmenttype:
          type: array
          description: |2
                    Schriftstücke
                    Schriftstücke können über diesen Aufruf nicht geändert bzw. ersetzt werden.
                    Neue Schriftstücke können bei Bedarf (zum Beispiel zum nachträglichen Hinzufügen
                    von Anhängen vor dem Postversand) über diesen Aufruf hinzugefügt werden.
          items:
            type: string
            format: binary
    MetadataReferenz:
      type: object
      properties:
        objaddress:
          type: string
        filename:
          type: string
        fileextension:
          type: string
        objclass:
          type: string
        contsize:
          type: string
        objcreatedby:
          $ref: '#/components/schemas/ArrayOfstring'
        objcreatedat:
          type: string
          format: date-time
        objchangedby:
          $ref: '#/components/schemas/ArrayOfstring'
        objmodifiedat:
          type: string
          format: date-time
...
    UpdateIncomingAntwortDTO:
      type: object
      properties:
        objid:
          type: string
          description: Objekt-ID (COO-Adresse) des Objekts
          example: COO.1.2301.1.1042441
        objname:
          type: string
          description: Objektname des Objekts
          example: Eingangspost vom 21.10.2016 (0010 A20 011-4-0006-0001)
        giobjecttype:
          type: array
          description: |
            Liste der Objektadressen und Namen
            der Schriftstücke im Dokument.
          example:
            - objname: Beispielname
              objaddress: COO.2150.9151.1.1206000
          items:
            $ref: '#/components/schemas/Objektreferenz'
...
    CreateIncomingAntwortDTO:
      type: object
      properties:
        objid:
          type: string
          description: Objekt-ID (COO-Adresse) des erzeugten Objekts
        objname:
          type: string
          description: Objektname des erzeugten Objekts
        giobjecttype:
          type: array
          description: >
            Auflistung aller Objekte, auf die der Suchstring zutrifft.

            Rückgabe der Eigenschaften, die im Datentyp GIObjectType definiert
            sind.
          example:
            - objname: Beispielname
              objaddress: COO.2150.9151.1.1206000
          items:
            $ref: '#/components/schemas/Objektreferenz'
...
    UpdateIncomingRequestBodyDTO:
      required:
        - params
      type: object
      properties:
        params:
          $ref: '#/components/schemas/UpdateIncomingAnfrageDTO'
        giattachmenttype:
          type: array
          description: |2
                                Schriftstücke
                                Schriftstücke können über diesen Aufruf nicht geändert bzw. ersetzt werden.
                                Neue Schriftstücke können bei Bedarf (zum Beispiel zum nachträglichen Hinzufügen von Anhängen
                                vor dem Postversand) über diesen Aufruf hinzugefügt werden.
          items:
            type: string
            format: binary
...
    ReadIncomingAntwortDTO:
      type: object
      properties:
        shortname:
          type: string
          description: Titel
          example: 2016 Anträge Firma XY
        filesubj:
          type: string
          description: |2
                      Betreff (mehrzeilig)
                      Mehrzeilige Felder werden mit Zeilenumbrüchen oder HTML Kodierung &#10; in Abfragen unterstützt.
          example: Alle Anträge der Firma XY aus dem Jahr 2016
        objterms:
          type: string
          description: >-
            Schlagworte, Angabe der Schlagworte getrennt mit „;“. Kein Semikola
            am Ende angeben.
          example: Firma XY; Anträge; Anträge 2016
        accdef:
          type: string
          description: >-
            Zugriffsdefinition

            Eindeutiger Name der Zugriffsdefinition. z.B.:

            • Zugriffsdefinition für Schriftgutobjekte (allgemein lesbar)

            • Zugriffsdefinition für Schriftgutobjekte der Organisationseinheit

            • Zugriffsdefinition für Schriftgutobjekte der Organisationseinheit
            und Zentralregistratur

            • Zugriffsdefinition für Schriftgutobjekte des Eigentümers

            Hinweis: Hierbei kann die gewünschte Zugriffsdefinition der Akte
            gesetzt werden. Wenn kein Wert übergeben wird, wird die auf dem
            Aktenplaneintrag hinterlegte Zugriffsdefinition initialisiert.
          example: Zugriffsdefinition für Schriftgutobjekte (allgemein lesbar)
        objname:
          type: string
          description: Eingangskennzeichen
          example: Eingangspost vom 21.10.2016 (0010 A20 011-4-0006-0001)
        referrednumber:
          type: string
          description: Objekt-ID (COO-Adresse) der zugehörigen Vorgangs
          example: COO.1.2301.1.1042432
        foreignnr:
          type: string
          description: Fremdes Geschäftszeichen
          example: A2016-10-2016-Landeshauptstadt-München
        delivery:
          type: string
          description: Termin (Datum/Uhrzeit)
          format: date-time
          example: '2016-10-26T18:16:29+01:00'
        documentremarks:
          type: string
          description: Dokumentbezogene Hinweise
          example: Antrag auf Baugenehmigung
        incattachments:
          type: string
          description: Sonstige Anlagen
          example: Ausfüllhilfe Antrag auf Baugenehmigung
        gimetadatatype:
          type: array
          description: Schriftstücke
          example:
            objaddress: COO.1.2301.1.1041875
            filename: Antrag auf Baugenehmigung
            fileextension: pdf
            objclass: PDF-Dokument
            contsize: '243'
            objcreatedby:
              string:
                - musterfraum
            objcreatedat: '2018-07-23T08:59:52+01:00'
            objchangedby:
              string:
                - mustermannm
            objmodifiedat: '2018-07-23T09:23:52+01:00'
          items:
            $ref: '#/components/schemas/MetadataReferenz'
...
    Objektreferenz:
      type: object
      properties:
        objname:
          type: string
        objaddress:
          type: string
      description: |-
        Vorgänge
        Liste der Vorgänge der Akte
      example:
        - objname: Beispielname
          objaddress: COO.2150.9151.1.1206000
...
  securitySchemes:
    basicAuthSecurityDefintion:
      type: http
      description: >-
        This API uses Basic Auth flow. [More
        info](https://api.example.com/docs/auth)
      scheme: basic