# v1/eakte-api-v1/eakte-api-v1-spring-boot-server-stubs/src-generated/src/main/java/de/muenchen/itm/eakte/schnittstelle/rest_v2/server_stubs/controllers/IncomingsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.21.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.controllers;

import de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.model.CreateIncomingAntwortDTO;
import de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.model.CreateIncomingBasisAnfrageDTO;
import de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.model.DmsErrorResponse;
import org.springframework.lang.Nullable;
import de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.model.UpdateIncomingAnfrageDTO;
import de.muenchen.itm.eakte.schnittstelle.rest_v2.server_stubs.model.UpdateIncomingAntwortDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;
import jakarta.servlet.http.HttpServletRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.21.0")
@Validated
@Controller
@Tag(name = "/incomings", description = "the /incomings API")
public interface IncomingsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_INCOMING = "/incomings";
    /**
     * POST /incomings : Eingangsdokument anlegen
     * Neues Eingangsdokument erstellen unter dem angegebenen Vorgang (optional inklusive Schriftstücke).\\ Wenn das Eingangsdokument keinem Vorgang zugeordnet ist (referrednumber &#x3D;&#x3D; NULL),  wird eine neue Aktivität „Eingang“ im Arbeitsvorrat erzeugt.  Standardmäßig wird diese dem Benutzer mit dessen Standardrolle in dessen Arbeitsvorrat vorgeschrieben.\\ Sollte jedoch der Parameter useOE &#x3D;&#x3D; true sein,  so soll die Aktivität „Eingang“ der OE (aus der Standardrolle des Benutzers) vorgeschrieben werden –  in diesem Fall, soll bei der Vorschreibung kein Benutzer eingetragen werden.  Beim Anlegen von neuen Dokumenten wird die auf dem Vorgang gesetzte Zugriffsdefinition verwendet.  Anhand des Userlogins werden der Eigentümer (COOSYSTEM@1.1:objowner) und die Organisationseinheit (COOSYSTEM@1.1:objowngroud) gesetzt. 
     *
     * @param userlogin **Benutzer-Login** (Benutzer-Kontext für den SOAP-Aufruf)\\ Log-in-Name des Benutzerobjekts  (required)
     * @param params  (required)
     * @param xAnwendung **Aufrufende Fachanwendung**  Für jeden Aufruf soll ersichtlich sein,  welche Fachanwendung den Aufruf getätigt hat.  Dies soll es ermöglichen, Requests und Responses den entsprechenden Fachverfahren zuzuordnen.  Für diesen Zweck wird sowohl im Request,  als auch im Response ein Parameter „businessapp“ angegeben.  Dieser Parameter ist optional und kann mit einem Namen befüllt werden.  Sofern der Aufruf einen Wert enthält, wird dieser auch im Response zurückgegeben.  Somit ist ein nachträgliches Filtern von Aufrufen möglich.  (optional)
     * @param joboe COO-Adresse der zu verwendenden **Organisationseinheit** des angemeldeten Nutzers.\\ \\ Ohne Angabe der joboe und jobposition  wird die eingestellte Standardrolle und OE des Nutzers in der eAkte verwendet.\\ Wenn eine bestimmte jobposition angegeben wird, unter der dieser Aufruf ausgeführt werden soll  muss auch die entsprechende joboe des Nutzers mit angegeben werden.  Das kann z.B. notwendig sein, wenn bestimmte Funktionen nur durch Schriftgutverwalter oder Registratur ausgeführt werden dürfen.\\ Über die OE wird der Mandant zugeordnet und entsprechend die Berechtigungen geprüft.\\ Hinweis:\\ Die COO-Adresse für die joboe muss im Fachverfahren in einer Konfigurationsdatei hintelegt werden, da sie nicht per Schnittstelle ausgelesen werden kann und sich ggf. ändern kann.  Meistens ist die joboe innerhalb eines Fachbereichs für alle Benutzer gleich, so dass sie bei Änderungen einfach zu verwalten ist.  (optional)
     * @param jobposition Referenz der Stelle (**Rolle des Nutzers** in der eAkte). * Schriftgutverwaltung - DocumentManager * Sachbearbeitung - OfficialInCharge * Fachadministration - OpAdm * Registratur - Official * Sekretariat - Secretary * Leitung - Head * Archivar - Registrar * Vorlagenadministrator - TemplateAdministrator * Vorlagenverwaltung - TemplateManager  Es dürfen ausschließlich die o.g. genannten Stellen der LHM verwendet werden.\\ Bitte verwenden Sie die englische Bezeichnung.  Die Rolle Fachadministrator ist bei Schnittstellen nicht erlaubt!  (optional)
     * @param giattachmenttype Liste von Schriftstücken\\\\ Array im Format: COO-Adresse, Dateiname, Inhalt der Datei, Dateiendung, Größe der Datei (in KB), Betreff  Hinweis:\\\\ Schriftstücke können über diesen Aufruf nicht geändert bzw. ersetzt werden.\\\\ Neue Schriftstücke können bei Bedarf (zum Beispiel zum nachträglichen Hinzufügen von Anhängen vor dem Postversand) über diesen Aufruf hinzugefügt werden.  (optional)
     * @return OK (status code 200)
     *         or Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten. (status code 400)
     *         or Falscher oder fehlender technischer Nutzer. (status code 401)
     *         or Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten. (status code 500)
     */
    @Operation(
        operationId = "createIncoming",
        summary = "Eingangsdokument anlegen",
        description = "Neues Eingangsdokument erstellen unter dem angegebenen Vorgang (optional inklusive Schriftstücke).\\ Wenn das Eingangsdokument keinem Vorgang zugeordnet ist (referrednumber == NULL),  wird eine neue Aktivität „Eingang“ im Arbeitsvorrat erzeugt.  Standardmäßig wird diese dem Benutzer mit dessen Standardrolle in dessen Arbeitsvorrat vorgeschrieben.\\ Sollte jedoch der Parameter useOE == true sein,  so soll die Aktivität „Eingang“ der OE (aus der Standardrolle des Benutzers) vorgeschrieben werden –  in diesem Fall, soll bei der Vorschreibung kein Benutzer eingetragen werden.  Beim Anlegen von neuen Dokumenten wird die auf dem Vorgang gesetzte Zugriffsdefinition verwendet.  Anhand des Userlogins werden der Eigentümer (COOSYSTEM@1.1:objowner) und die Organisationseinheit (COOSYSTEM@1.1:objowngroud) gesetzt. ",
        tags = { "/incomings" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreateIncomingAntwortDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DmsErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Falscher oder fehlender technischer Nutzer."),
            @ApiResponse(responseCode = "500", description = "Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten.")
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = IncomingsApi.PATH_CREATE_INCOMING,
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    default ResponseEntity<CreateIncomingAntwortDTO> createIncoming(
        @NotNull @Parameter(name = "userlogin", description = "**Benutzer-Login** (Benutzer-Kontext für den SOAP-Aufruf)\\ Log-in-Name des Benutzerobjekts ", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "userlogin", required = true) String userlogin,
        @Parameter(name = "params", description = "", required = true) @Valid @RequestPart(value = "params", required = true) CreateIncomingBasisAnfrageDTO params,
        @Parameter(name = "x-anwendung", description = "**Aufrufende Fachanwendung**  Für jeden Aufruf soll ersichtlich sein,  welche Fachanwendung den Aufruf getätigt hat.  Dies soll es ermöglichen, Requests und Responses den entsprechenden Fachverfahren zuzuordnen.  Für diesen Zweck wird sowohl im Request,  als auch im Response ein Parameter „businessapp“ angegeben.  Dieser Parameter ist optional und kann mit einem Namen befüllt werden.  Sofern der Aufruf einen Wert enthält, wird dieser auch im Response zurückgegeben.  Somit ist ein nachträgliches Filtern von Aufrufen möglich. ", in = ParameterIn.HEADER) @RequestHeader(value = "x-anwendung", required = false) Optional<String> xAnwendung,
        @Parameter(name = "joboe", description = "COO-Adresse der zu verwendenden **Organisationseinheit** des angemeldeten Nutzers.\\ \\ Ohne Angabe der joboe und jobposition  wird die eingestellte Standardrolle und OE des Nutzers in der eAkte verwendet.\\ Wenn eine bestimmte jobposition angegeben wird, unter der dieser Aufruf ausgeführt werden soll  muss auch die entsprechende joboe des Nutzers mit angegeben werden.  Das kann z.B. notwendig sein, wenn bestimmte Funktionen nur durch Schriftgutverwalter oder Registratur ausgeführt werden dürfen.\\ Über die OE wird der Mandant zugeordnet und entsprechend die Berechtigungen geprüft.\\ Hinweis:\\ Die COO-Adresse für die joboe muss im Fachverfahren in einer Konfigurationsdatei hintelegt werden, da sie nicht per Schnittstelle ausgelesen werden kann und sich ggf. ändern kann.  Meistens ist die joboe innerhalb eines Fachbereichs für alle Benutzer gleich, so dass sie bei Änderungen einfach zu verwalten ist. ", in = ParameterIn.HEADER) @RequestHeader(value = "joboe", required = false) Optional<String> joboe,
        @Parameter(name = "jobposition", description = "Referenz der Stelle (**Rolle des Nutzers** in der eAkte). * Schriftgutverwaltung - DocumentManager * Sachbearbeitung - OfficialInCharge * Fachadministration - OpAdm * Registratur - Official * Sekretariat - Secretary * Leitung - Head * Archivar - Registrar * Vorlagenadministrator - TemplateAdministrator * Vorlagenverwaltung - TemplateManager  Es dürfen ausschließlich die o.g. genannten Stellen der LHM verwendet werden.\\ Bitte verwenden Sie die englische Bezeichnung.  Die Rolle Fachadministrator ist bei Schnittstellen nicht erlaubt! ", in = ParameterIn.HEADER) @RequestHeader(value = "jobposition", required = false) Optional<String> jobposition,
        @Parameter(name = "giattachmenttype", description = "Liste von Schriftstücken\\\\ Array im Format: COO-Adresse, Dateiname, Inhalt der Datei, Dateiendung, Größe der Datei (in KB), Betreff  Hinweis:\\\\ Schriftstücke können über diesen Aufruf nicht geändert bzw. ersetzt werden.\\\\ Neue Schriftstücke können bei Bedarf (zum Beispiel zum nachträglichen Hinzufügen von Anhängen vor dem Postversand) über diesen Aufruf hinzugefügt werden. ") @RequestPart(value = "giattachmenttype", required = false) List<MultipartFile> giattachmenttype,
        @Parameter(hidden = true) final HttpServletRequest servletRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"objname\" : \"objname\", \"giobjecttype\" : \"LHM-Sitzung 14.10.2016 COO.1.2301.1.1041875\", \"objid\" : \"objid\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"fehlerQuelle\" : \"INTERN\", \"text\" : \"text\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_INCOMING = "/incomings/{objaddress}";
    /**
     * PUT /incomings/{objaddress} : Eingang ändern
     * Die Werte werden mit den übergebenen Parametern überschrieben.  Bei keiner Übergabe bleiben die bestehenden Werte erhalten.\\ Ausnahme: Inhaltsobjekte können über diesen Aufruf nur hinzugefügt werden.\\ Dokumente können nur in bereits bestehenden Vorgängen angelegt werden. Bei jeder Änderung über die Schnittstelle wird in der eGov-Suite eine neue Version des Objekts erzeugt. Dies betrifft auch wiederholte Änderungen, die durch einen Benutzer durchgeführt werden. 
     *
     * @param userlogin **Benutzer-Login** (Benutzer-Kontext für den SOAP-Aufruf)\\ Log-in-Name des Benutzerobjekts  (required)
     * @param objaddress Objekt-ID (COO-Adresse) des Objekts (required)
     * @param params  (required)
     * @param xAnwendung **Aufrufende Fachanwendung**  Für jeden Aufruf soll ersichtlich sein,  welche Fachanwendung den Aufruf getätigt hat.  Dies soll es ermöglichen, Requests und Responses den entsprechenden Fachverfahren zuzuordnen.  Für diesen Zweck wird sowohl im Request,  als auch im Response ein Parameter „businessapp“ angegeben.  Dieser Parameter ist optional und kann mit einem Namen befüllt werden.  Sofern der Aufruf einen Wert enthält, wird dieser auch im Response zurückgegeben.  Somit ist ein nachträgliches Filtern von Aufrufen möglich.  (optional)
     * @param joboe COO-Adresse der zu verwendenden **Organisationseinheit** des angemeldeten Nutzers.\\ \\ Ohne Angabe der joboe und jobposition  wird die eingestellte Standardrolle und OE des Nutzers in der eAkte verwendet.\\ Wenn eine bestimmte jobposition angegeben wird, unter der dieser Aufruf ausgeführt werden soll  muss auch die entsprechende joboe des Nutzers mit angegeben werden.  Das kann z.B. notwendig sein, wenn bestimmte Funktionen nur durch Schriftgutverwalter oder Registratur ausgeführt werden dürfen.\\ Über die OE wird der Mandant zugeordnet und entsprechend die Berechtigungen geprüft.\\ Hinweis:\\ Die COO-Adresse für die joboe muss im Fachverfahren in einer Konfigurationsdatei hintelegt werden, da sie nicht per Schnittstelle ausgelesen werden kann und sich ggf. ändern kann.  Meistens ist die joboe innerhalb eines Fachbereichs für alle Benutzer gleich, so dass sie bei Änderungen einfach zu verwalten ist.  (optional)
     * @param jobposition Referenz der Stelle (**Rolle des Nutzers** in der eAkte). * Schriftgutverwaltung - DocumentManager * Sachbearbeitung - OfficialInCharge * Fachadministration - OpAdm * Registratur - Official * Sekretariat - Secretary * Leitung - Head * Archivar - Registrar * Vorlagenadministrator - TemplateAdministrator * Vorlagenverwaltung - TemplateManager  Es dürfen ausschließlich die o.g. genannten Stellen der LHM verwendet werden.\\ Bitte verwenden Sie die englische Bezeichnung.  Die Rolle Fachadministrator ist bei Schnittstellen nicht erlaubt!  (optional)
     * @param giattachmenttype Liste der Schriftstücke\\\\ Schriftstücke können über diesen Aufruf nicht geändert bzw. ersetzt werden. Neue Schriftstücke können bei Bedarf (zum Beispiel zum nachträglichen Hinzufügen von Anhängen vor dem Postversand) über diesen Aufruf hinzugefügt werden.  (optional)
     * @return OK (status code 200)
     *         or Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten. (status code 400)
     *         or Falscher oder fehlender technischer Nutzer. (status code 401)
     *         or Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten. (status code 500)
     */
    @Operation(
        operationId = "updateIncoming",
        summary = "Eingang ändern",
        description = "Die Werte werden mit den übergebenen Parametern überschrieben.  Bei keiner Übergabe bleiben die bestehenden Werte erhalten.\\ Ausnahme: Inhaltsobjekte können über diesen Aufruf nur hinzugefügt werden.\\ Dokumente können nur in bereits bestehenden Vorgängen angelegt werden. Bei jeder Änderung über die Schnittstelle wird in der eGov-Suite eine neue Version des Objekts erzeugt. Dies betrifft auch wiederholte Änderungen, die durch einen Benutzer durchgeführt werden. ",
        tags = { "/incomings" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UpdateIncomingAntwortDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Aufruf ans DMS ist gescheitert. Im Body sind Details enthalten.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DmsErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Falscher oder fehlender technischer Nutzer."),
            @ApiResponse(responseCode = "500", description = "Ein unerwarteter Fehler innerhalb der EAI ist aufgetreten.")
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = IncomingsApi.PATH_UPDATE_INCOMING,
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    default ResponseEntity<UpdateIncomingAntwortDTO> updateIncoming(
        @NotNull @Parameter(name = "userlogin", description = "**Benutzer-Login** (Benutzer-Kontext für den SOAP-Aufruf)\\ Log-in-Name des Benutzerobjekts ", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "userlogin", required = true) String userlogin,
        @Parameter(name = "objaddress", description = "Objekt-ID (COO-Adresse) des Objekts", required = true, in = ParameterIn.PATH) @PathVariable("objaddress") String objaddress,
        @Parameter(name = "params", description = "", required = true) @Valid @RequestPart(value = "params", required = true) UpdateIncomingAnfrageDTO params,
        @Parameter(name = "x-anwendung", description = "**Aufrufende Fachanwendung**  Für jeden Aufruf soll ersichtlich sein,  welche Fachanwendung den Aufruf getätigt hat.  Dies soll es ermöglichen, Requests und Responses den entsprechenden Fachverfahren zuzuordnen.  Für diesen Zweck wird sowohl im Request,  als auch im Response ein Parameter „businessapp“ angegeben.  Dieser Parameter ist optional und kann mit einem Namen befüllt werden.  Sofern der Aufruf einen Wert enthält, wird dieser auch im Response zurückgegeben.  Somit ist ein nachträgliches Filtern von Aufrufen möglich. ", in = ParameterIn.HEADER) @RequestHeader(value = "x-anwendung", required = false) Optional<String> xAnwendung,
        @Parameter(name = "joboe", description = "COO-Adresse der zu verwendenden **Organisationseinheit** des angemeldeten Nutzers.\\ \\ Ohne Angabe der joboe und jobposition  wird die eingestellte Standardrolle und OE des Nutzers in der eAkte verwendet.\\ Wenn eine bestimmte jobposition angegeben wird, unter der dieser Aufruf ausgeführt werden soll  muss auch die entsprechende joboe des Nutzers mit angegeben werden.  Das kann z.B. notwendig sein, wenn bestimmte Funktionen nur durch Schriftgutverwalter oder Registratur ausgeführt werden dürfen.\\ Über die OE wird der Mandant zugeordnet und entsprechend die Berechtigungen geprüft.\\ Hinweis:\\ Die COO-Adresse für die joboe muss im Fachverfahren in einer Konfigurationsdatei hintelegt werden, da sie nicht per Schnittstelle ausgelesen werden kann und sich ggf. ändern kann.  Meistens ist die joboe innerhalb eines Fachbereichs für alle Benutzer gleich, so dass sie bei Änderungen einfach zu verwalten ist. ", in = ParameterIn.HEADER) @RequestHeader(value = "joboe", required = false) Optional<String> joboe,
        @Parameter(name = "jobposition", description = "Referenz der Stelle (**Rolle des Nutzers** in der eAkte). * Schriftgutverwaltung - DocumentManager * Sachbearbeitung - OfficialInCharge * Fachadministration - OpAdm * Registratur - Official * Sekretariat - Secretary * Leitung - Head * Archivar - Registrar * Vorlagenadministrator - TemplateAdministrator * Vorlagenverwaltung - TemplateManager  Es dürfen ausschließlich die o.g. genannten Stellen der LHM verwendet werden.\\ Bitte verwenden Sie die englische Bezeichnung.  Die Rolle Fachadministrator ist bei Schnittstellen nicht erlaubt! ", in = ParameterIn.HEADER) @RequestHeader(value = "jobposition", required = false) Optional<String> jobposition,
        @Parameter(name = "giattachmenttype", description = "Liste der Schriftstücke\\\\ Schriftstücke können über diesen Aufruf nicht geändert bzw. ersetzt werden. Neue Schriftstücke können bei Bedarf (zum Beispiel zum nachträglichen Hinzufügen von Anhängen vor dem Postversand) über diesen Aufruf hinzugefügt werden. ") @RequestPart(value = "giattachmenttype", required = false) List<MultipartFile> giattachmenttype,
        @Parameter(hidden = true) final HttpServletRequest servletRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"objname\" : \"Eingangspost vom 28.05.2026 (0010-11/2-0006)\", \"giobjecttype\" : [ [ { \"name\" : \"Beispielname\", \"id\" : \"COO.2150.9151.1.1206000\" } ], [ { \"name\" : \"Beispielname\", \"id\" : \"COO.2150.9151.1.1206000\" } ] ], \"objid\" : \"COO.1.2301.1.1042441\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"fehlerQuelle\" : \"INTERN\", \"text\" : \"text\", \"status\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}